+++ db/migrate/20230720145221_add2ri_to_projects.rb
@@ -1,0 +1,5 @@
+class Add2riToProjects < ActiveRecord::Migration[7.0]
+  def change
+    add_column :projects, :two_requires_improvement, :boolean, default: false
+  end
+end