+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.0].define(version: 2023_07_18_141059) do
+ActiveRecord::Schema[7.0].define(version: 2023_07_20_145221) do
   create_table "contacts", id: :uuid, default: -> { "newid()" }, force: :cascade do |t|
     t.uuid "project_id"
     t.string "name", null: false
@@ -204,6 +204,7 @@ ActiveRecord::Schema[7.0].define(version: 2023_07_18_141059) do
     t.uuid "funding_agreement_contact_id"
     t.integer "outgoing_trust_ukprn"
     t.string "team"
+    t.boolean "two_requires_improvement", default: false
     t.index ["assigned_to_id"], name: "index_projects_on_assigned_to_id"
     t.index ["caseworker_id"], name: "index_projects_on_caseworker_id"
     t.index ["regional_delivery_officer_id"], name: "index_projects_on_regional_delivery_officer_id"